+++ src/Entity/Log/HistorizableTrait.php
@@ -3,12 +3,16 @@
 namespace Smart\SonataBundle\Entity\Log;
 
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\PropertyAccess\PropertyAccess;
+use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 
 /**
  * @author Mathieu Ducrot <mathieu.ducrot@smartbooster.io>
  */
 trait HistorizableTrait
 {
+    private ?PropertyAccessorInterface $propertyAccess = null;
+
     /**
      * @ORM\Column(type="json", nullable=true)
      */
@@ -34,5 +38,36 @@ trait HistorizableTrait
         array_unshift($this->history, $history);
 
         return $this;
+    }
+
+    public function getDataForHistoryDiff(): array
+    {
+        $propertyAccessor = $this->getPropertyAccessor();
+
+        $toReturn = [];
+        foreach ($this->getAttributsForHistoryDiff() as $attribut) {
+            $toReturn[$attribut] = $propertyAccessor->getValue($this, $attribut);
+        }
+
+        return $toReturn;
+    }
+
+    /**
+     * Redefine in entity
+     */
+    public function getAttributsForHistoryDiff(): array
+    {
+        return [];
+    }
+
+    private function getPropertyAccessor(): PropertyAccessorInterface
+    {
+        if (!$this->propertyAccess instanceof PropertyAccessorInterface) {
+            $this->propertyAccess = PropertyAccess::createPropertyAccessorBuilder()
+                ->enableExceptionOnInvalidIndex()
+                ->getPropertyAccessor();
+        }
+
+        return $this->propertyAccess;
     }
 }